Move reporting logic to core module (#2754)
* Move console reports to core module

* Simplify cli test cases

* Move to console package

* Move core reporting logic from cli to core module
