+++ detekt-cli/build.gradle.kts
@@ -5,9 +5,6 @@ application {
 dependencies {
     implementation(kotlin("stdlib-jdk8"))
     implementation(project(":detekt-core"))
-    implementation(project(":detekt-report-html"))
-    implementation(project(":detekt-report-txt"))
-    implementation(project(":detekt-report-xml"))
     runtimeOnly(project(":detekt-rules"))
     implementation("com.beust:jcommander:${Versions.JCOMMANDER}")
