+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/IssueExtension.kt
@@ -1,23 +1,19 @@
 package io.gitlab.arturbosch.detekt.cli
 
 import io.gitlab.arturbosch.detekt.api.Config
-import io.gitlab.arturbosch.detekt.api.CorrectableCodeSmell
 import io.gitlab.arturbosch.detekt.api.Detektion
 import io.gitlab.arturbosch.detekt.api.Finding
 import io.gitlab.arturbosch.detekt.api.RuleId
 import io.gitlab.arturbosch.detekt.api.RuleSetId
+import io.gitlab.arturbosch.detekt.core.reporting.BUILD
+import io.gitlab.arturbosch.detekt.core.reporting.filterAutoCorrectedIssues
 import org.jetbrains.kotlin.com.intellij.openapi.util.Key
-import java.util.HashMap
 
 private val WEIGHTED_ISSUES_COUNT_KEY = Key.create<Int>("WEIGHTED_ISSUES_COUNT")
-private const val BUILD = "build"
 private const val WEIGHTS = "weights"
 private const val MAX_ISSUES = "maxIssues"
-private const val EXCLUDE_CORRECTABLE = "excludeCorrectable"
 
 fun Config.maxIssues(): Int = subConfig(BUILD).valueOrDefault(MAX_ISSUES, -1)
-
-fun Config.excludeCorrectable(): Boolean = subConfig(BUILD).valueOrDefault(EXCLUDE_CORRECTABLE, false)
 
 fun Int.isValidAndSmallerOrEqual(amount: Int): Boolean =
     !(this == 0 && amount == 0) && this != -1 && this <= amount
@@ -43,27 +39,6 @@ fun Detektion.getOrComputeWeightedAmountOfIssues(config: Config): Int {
     val amount = smells.sumBy { it.weighted() }
     this.addData(WEIGHTED_ISSUES_COUNT_KEY, amount)
     return amount
-}
-
-fun Detektion.filterEmptyIssues(config: Config): Map<RuleSetId, List<Finding>> {
-    return this
-        .filterAutoCorrectedIssues(config)
-        .filter { it.value.isNotEmpty() }
-}
-
-fun Detektion.filterAutoCorrectedIssues(config: Config): Map<RuleSetId, List<Finding>> {
-    if (!config.excludeCorrectable()) {
-        return findings
-    }
-    val filteredFindings = HashMap<RuleSetId, List<Finding>>()
-    findings.forEach { (ruleSetId, findingsList) ->
-        val newFindingsList = findingsList.filter { finding ->
-            val correctableCodeSmell = finding as? CorrectableCodeSmell
-            correctableCodeSmell == null || !correctableCodeSmell.autoCorrectEnabled
-        }
-        filteredFindings[ruleSetId] = newFindingsList
-    }
-    return filteredFindings
 }
 
 private fun Config.weightsConfig(): Config = subConfig(BUILD).subConfig(WEIGHTS)
