+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/Junk.kt
@@ -2,5 +2,3 @@ package io.gitlab.arturbosch.detekt.cli
 
 const val SEPARATOR_COMMA = ","
 const val SEPARATOR_SEMICOLON = ";"
-
-val IS_WINDOWS = System.getProperty("os.name").contains("Windows")
