--- detekt-cli/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.ConsoleReport
@@ -1,5 +1,0 @@
-io.gitlab.arturbosch.detekt.cli.console.ComplexityReport
-io.gitlab.arturbosch.detekt.cli.console.FindingsReport
-io.gitlab.arturbosch.detekt.cli.console.FileBasedFindingsReport
-io.gitlab.arturbosch.detekt.cli.console.NotificationReport
-io.gitlab.arturbosch.detekt.cli.console.ProjectStatisticsReport
