--- detekt-cli/src/test/kotlin/io/gitlab/arturbosch/detekt/cli/console/ResourceReader.kt
@@ -1,9 +1,0 @@
-package io.gitlab.arturbosch.detekt.cli.console
-
-import io.github.detekt.test.utils.resourceAsPath
-import java.nio.file.Files
-
-internal fun readResource(filename: String): String {
-    val path = resourceAsPath(filename)
-    return Files.readAllLines(path).joinToString("\n") + "\n"
-}
