--- detekt-cli/src/test/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.OutputReport
@@ -1,1 +1,0 @@
-io.gitlab.arturbosch.detekt.cli.out.TestOutputReport
