+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/reporting/OutputFacade.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.cli
+package io.gitlab.arturbosch.detekt.core.reporting
 
 import io.gitlab.arturbosch.detekt.api.ConsoleReport
 import io.gitlab.arturbosch.detekt.api.Detektion
@@ -7,12 +7,12 @@ import io.gitlab.arturbosch.detekt.api.internal.SimpleNotification
 import io.gitlab.arturbosch.detekt.core.ProcessingSettings
 
 class OutputFacade(
-    arguments: CliArgs,
+    reportPaths: List<ReportPath>,
     private val result: Detektion,
     private val settings: ProcessingSettings
 ) {
 
-    private val reportPaths: Map<String, ReportPath> = arguments.reportPaths.associateBy { it.kind }
+    private val reports: Map<String, ReportPath> = reportPaths.associateBy { it.kind }
 
     fun run() {
         OutputReportLocator(settings)
@@ -24,7 +24,7 @@ class OutputFacade(
     }
 
     private fun handleOutputReport(report: OutputReport) {
-        val filePath = reportPaths[report.id]?.path
+        val filePath = reports[report.id]?.path
         if (filePath != null) {
             report.write(filePath, result)
             result.add(SimpleNotification("Successfully generated ${report.name} at $filePath"))
