+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/reporting/ReportLocator.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.cli
+package io.gitlab.arturbosch.detekt.core.reporting
 
 import io.gitlab.arturbosch.detekt.api.ConsoleReport
 import io.gitlab.arturbosch.detekt.api.Extension
