+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/reporting/console/ComplexityReport.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.cli.console
+package io.gitlab.arturbosch.detekt.core.reporting.console
 
 import io.github.detekt.metrics.ComplexityReportGenerator
 import io.gitlab.arturbosch.detekt.api.ConsoleReport
