+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/reporting/console/FileBasedFindingsReport.kt
@@ -1,12 +1,13 @@
-package io.gitlab.arturbosch.detekt.cli.console
+package io.gitlab.arturbosch.detekt.core.reporting.console
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.ConsoleReport
 import io.gitlab.arturbosch.detekt.api.Detektion
 import io.gitlab.arturbosch.detekt.api.SingleAssign
-import io.gitlab.arturbosch.detekt.cli.filterEmptyIssues
+import io.gitlab.arturbosch.detekt.core.reporting.filterEmptyIssues
+import io.gitlab.arturbosch.detekt.core.reporting.printFindings
 
-class FindingsReport : ConsoleReport() {
+class FileBasedFindingsReport : ConsoleReport() {
 
     private var config: Config by SingleAssign()
 
@@ -22,6 +23,9 @@ class FindingsReport : ConsoleReport() {
             return null
         }
 
-        return printFindings(findings)
+        val findingsPerFile = findings.values
+            .flatten()
+            .groupBy { it.entity.location.file }
+        return printFindings(findingsPerFile)
     }
 }
