+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/reporting/console/ProjectStatisticsReport.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.cli.console
+package io.gitlab.arturbosch.detekt.core.reporting.console
 
 import io.gitlab.arturbosch.detekt.api.ConsoleReport
 import io.gitlab.arturbosch.detekt.api.Detektion
