+++ detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/reporting/OutputFacadeSpec.kt
@@ -1,10 +1,11 @@
-package io.gitlab.arturbosch.detekt.cli
+package io.gitlab.arturbosch.detekt.core.reporting
 
 import io.github.detekt.report.html.HtmlOutputReport
 import io.github.detekt.report.txt.TxtOutputReport
 import io.github.detekt.report.xml.XmlOutputReport
 import io.github.detekt.test.utils.StringPrintStream
-import io.github.detekt.test.utils.resource
+import io.github.detekt.test.utils.createTempFileForTest
+import io.github.detekt.test.utils.resourceAsPath
 import io.gitlab.arturbosch.detekt.core.DetektResult
 import io.gitlab.arturbosch.detekt.test.createFinding
 import io.gitlab.arturbosch.detekt.test.createProcessingSettings
@@ -12,46 +13,40 @@ import org.assertj.core.api.Assertions.assertThat
 import org.jetbrains.kotlin.utils.closeQuietly
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
-import java.io.File
 import java.nio.file.Path
-import java.nio.file.Paths
 
 internal class OutputFacadeSpec : Spek({
 
     describe("Running the output facade with multiple reports") {
 
         val printStream = StringPrintStream()
-        val inputPath: Path = Paths.get(resource("/cases"))
-        lateinit var plainOutputPath: File
-        lateinit var htmlOutputPath: File
-        lateinit var xmlOutputPath: File
+        val inputPath: Path = resourceAsPath("/cases")
+        lateinit var plainOutputPath: Path
+        lateinit var htmlOutputPath: Path
+        lateinit var xmlOutputPath: Path
 
         val defaultDetektion = DetektResult(mapOf(Pair("Key", listOf(createFinding()))))
         val defaultSettings = createProcessingSettings(inputPath, outPrinter = printStream)
 
-        lateinit var cliArgs: CliArgs
+        lateinit var reportPaths: List<ReportPath>
 
         beforeEachTest {
-            plainOutputPath = File.createTempFile("detekt", ".txt")
-            htmlOutputPath = File.createTempFile("detekt", ".html")
-            xmlOutputPath = File.createTempFile("detekt", ".xml")
-            cliArgs = createCliArgs(
-                "--input", inputPath.toString(),
-                "--report", "xml:$xmlOutputPath",
-                "--report", "txt:$plainOutputPath",
-                "--report", "html:$htmlOutputPath"
-            )
+            plainOutputPath = createTempFileForTest("detekt", ".txt")
+            htmlOutputPath = createTempFileForTest("detekt", ".html")
+            xmlOutputPath = createTempFileForTest("detekt", ".xml")
+            reportPaths = listOf(
+                "xml:$xmlOutputPath",
+                "txt:$plainOutputPath",
+                "html:$htmlOutputPath"
+            ).map { ReportPath.from(it) }
         }
 
-        afterEachTest {
-            plainOutputPath.delete()
-            htmlOutputPath.delete()
-            xmlOutputPath.delete()
+        afterGroup {
             closeQuietly(defaultSettings)
         }
 
         it("creates all output files") {
-            val subject = OutputFacade(cliArgs, defaultDetektion, defaultSettings)
+            val subject = OutputFacade(reportPaths, defaultDetektion, defaultSettings)
 
             subject.run()
