+++ detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/reporting/ReportPathSpec.kt
@@ -1,5 +1,6 @@
-package io.gitlab.arturbosch.detekt.cli
+package io.gitlab.arturbosch.detekt.core.reporting
 
+import io.gitlab.arturbosch.detekt.core.IS_WINDOWS
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.assertj.core.api.Assertions.assertThatIllegalStateException
