+++ detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/reporting/console/NotificationReportSpec.kt
@@ -1,7 +1,7 @@
-package io.gitlab.arturbosch.detekt.cli.console
+package io.gitlab.arturbosch.detekt.core.reporting.console
 
 import io.github.detekt.test.utils.resourceAsPath
-import io.gitlab.arturbosch.detekt.cli.createNotification
+import io.gitlab.arturbosch.detekt.core.ModificationNotification
 import io.gitlab.arturbosch.detekt.test.TestDetektion
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
@@ -14,7 +14,7 @@ class NotificationReportSpec : Spek({
     describe("notification report") {
 
         it("reports two notifications") {
-            val path = resourceAsPath("empty.txt")
+            val path = resourceAsPath("/reporting/empty.txt")
             val detektion = object : TestDetektion() {
                 override val notifications = listOf(createNotification(), createNotification())
             }
@@ -29,3 +29,5 @@ class NotificationReportSpec : Spek({
 })
 
 private val LN = System.lineSeparator()
+
+private fun createNotification() = ModificationNotification(resourceAsPath("/reporting/empty.txt"))
