+++ detekt-core/src/test/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.OutputReport
@@ -1,0 +1,1 @@
+io.gitlab.arturbosch.detekt.core.reporting.TestOutputReport
