+++ detekt-core/src/test/resources/reporting/complexity-report.txt
