+++ detekt-core/src/test/resources/reporting/empty.txt
