+++ detekt-core/src/test/resources/reporting/findings-report.txt
