+++ detekt-core/src/test/resources/reporting/grouped-findings-report.txt
