+++ detekt-sample-extensions/src/test/kotlin/io/gitlab/arturbosch/detekt/sample/extensions/SupportConfigValidationSpec.kt
@@ -1,11 +1,10 @@
 package io.gitlab.arturbosch.detekt.sample.extensions
 
-import io.gitlab.arturbosch.detekt.cli.CliArgs
-import io.gitlab.arturbosch.detekt.cli.config.InvalidConfig
-import io.gitlab.arturbosch.detekt.cli.console.red
-import io.gitlab.arturbosch.detekt.cli.runners.Runner
 import io.github.detekt.test.utils.NullPrintStream
 import io.github.detekt.test.utils.createTempDirectoryForTest
+import io.gitlab.arturbosch.detekt.cli.CliArgs
+import io.gitlab.arturbosch.detekt.cli.config.InvalidConfig
+import io.gitlab.arturbosch.detekt.cli.runners.Runner
 import org.assertj.core.api.Assertions.assertThatCode
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
@@ -29,7 +28,7 @@ class SupportConfigValidationSpec : Spek({
 
                     assertThatCode { Runner(args, NullPrintStream(), NullPrintStream()).execute() }
                         .isInstanceOf(InvalidConfig::class.java)
-                        .hasMessage("Run failed with 1 invalid config property.".red())
+                        .hasMessageContaining("Run failed with 1 invalid config property.")
                 }
             }
         }
